Refactor layout components and update imports
